- Ghosts model, view and tests implemented - characters initial position changed - kapman's speed increased - all characters  positions and speed will be qreal instead of int
svn path=/trunk/playground/games/kapman/; revision=744414

+++ tests/kapmantest.cpp
@@ -5,65 +5,65 @@ void KapmanTest::testKapman() {
 
 	Kapman kap;
 	
-	QCOMPARE(kap.getX(), 0);
-	QCOMPARE(kap.getY(), 0);
-	QCOMPARE(kap.getXSpeed(), 0);
-	QCOMPARE(kap.getYSpeed(), 0);
+	QCOMPARE(kap.getX(), 0.0);
+	QCOMPARE(kap.getY(), 0.0);
+	QCOMPARE(kap.getXSpeed(), 0.0);
+	QCOMPARE(kap.getYSpeed(), 0.0);
 }
 
 void KapmanTest::testGoUp() {
 
 	Kapman kap;
-	int kapSpeed = kap.getSPEED();
+	qreal kapSpeed = kap.getSPEED();
 
 	kap.goUp();
 
-	QCOMPARE(kap.getXSpeed(), 0);	
+	QCOMPARE(kap.getXSpeed(), 0.0);	
 	QCOMPARE(kap.getYSpeed(), -kapSpeed);
 }
 
 void KapmanTest::testGoDown() {
 
 	Kapman kap;
-	int kapSpeed = kap.getSPEED();
+	qreal kapSpeed = kap.getSPEED();
 
 	kap.goDown();
 
-	QCOMPARE(kap.getXSpeed(), 0);	
+	QCOMPARE(kap.getXSpeed(), 0.0);	
 	QCOMPARE(kap.getYSpeed(), kapSpeed);
 }
 
 void KapmanTest::testGoRight() {
 
 	Kapman kap;
-	int kapSpeed = kap.getSPEED();
+	qreal kapSpeed = kap.getSPEED();
 
 	kap.goRight();
 
 	QCOMPARE(kap.getXSpeed(), kapSpeed);	
-	QCOMPARE(kap.getYSpeed(), 0);
+	QCOMPARE(kap.getYSpeed(), 0.0);
 }
 
 void KapmanTest::testGoLeft() {
 
 	Kapman kap;
-	int kapSpeed = kap.getSPEED();
+	qreal kapSpeed = kap.getSPEED();
 
 	kap.goLeft();
 
 	QCOMPARE(kap.getXSpeed(), -kapSpeed);	
-	QCOMPARE(kap.getYSpeed(), 0);
+	QCOMPARE(kap.getYSpeed(), 0.0);
 }
 
 void KapmanTest::testMove() {	
 	
 	Kapman * kap = new Kapman();
 	
-	int kapX = kap->getX();	// The x coordinate before calling move()
-	int kapY = kap->getY();	// The y coordinate before calling move()
+	qreal kapX = kap->getX();	// The x coordinate before calling move()
+	qreal kapY = kap->getY();	// The y coordinate before calling move()
 	
 	// Spy on moved signal
-	QSignalSpy viewUpdate(kap, SIGNAL(moved(int, int)));
+	QSignalSpy viewUpdate(kap, SIGNAL(moved(qreal, qreal)));
 	
 	kap->goUp();
 	kap->move();
@@ -88,14 +88,14 @@ void KapmanTest::testMove() {
 	// results of the spy viewUpdate
 	QCOMPARE(viewUpdate.count(), 3);
 	
-	QCOMPARE(viewUpdate.at(0).at(0).toInt(), kapX);
-	QCOMPARE(viewUpdate.at(0).at(1).toInt(), kapY - kap->getSPEED());
+	QCOMPARE(viewUpdate.at(0).at(0).toDouble(), kapX);
+	QCOMPARE(viewUpdate.at(0).at(1).toDouble(), kapY - kap->getSPEED());
 	
-	QCOMPARE(viewUpdate.at(1).at(0).toInt(), kapX);
-	QCOMPARE(viewUpdate.at(1).at(1).toInt(), kapY - 2*kap->getSPEED());
+	QCOMPARE(viewUpdate.at(1).at(0).toDouble(), kapX);
+	QCOMPARE(viewUpdate.at(1).at(1).toDouble(), kapY - 2*kap->getSPEED());
 	
-	QCOMPARE(viewUpdate.at(2).at(0).toInt(), kapX + kap->getSPEED());
-	QCOMPARE(viewUpdate.at(2).at(1).toInt(), kapY - 2*kap->getSPEED());
+	QCOMPARE(viewUpdate.at(2).at(0).toDouble(), kapX + kap->getSPEED());
+	QCOMPARE(viewUpdate.at(2).at(1).toDouble(), kapY - 2*kap->getSPEED());
 }
